refactor: Extract common test methods to own project
