+++ WarmCorners.sln
@@ -24,6 +24,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Unit", "Unit", "{57D3B586-2
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration", "{A698F53C-280F-47E6-9910-F5B8DD3BF56B}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{5637E214-6C55-4085-8A38-F41E29EA4122}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WarmCorners.Tests.Common", "tests\WarmCorners.Tests.Common\WarmCorners.Tests.Common.csproj", "{14B0C460-A362-47B4-A050-8B332DB79E2E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -66,6 +70,10 @@ Global
 		{FC12A768-F811-4CFD-B623-0AE749BA1354}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FC12A768-F811-4CFD-B623-0AE749BA1354}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FC12A768-F811-4CFD-B623-0AE749BA1354}.Release|Any CPU.Build.0 = Release|Any CPU
+		{14B0C460-A362-47B4-A050-8B332DB79E2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14B0C460-A362-47B4-A050-8B332DB79E2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14B0C460-A362-47B4-A050-8B332DB79E2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14B0C460-A362-47B4-A050-8B332DB79E2E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{57D3B586-2C49-405F-8D79-C8418C0B3DF1} = {E94B9BC6-2E64-4808-A710-BD8E3975D11D}
@@ -75,5 +83,7 @@ Global
 		{E2D1CC3F-59CE-4D50-8FBD-BAAE665DF0CC} = {57D3B586-2C49-405F-8D79-C8418C0B3DF1}
 		{A698F53C-280F-47E6-9910-F5B8DD3BF56B} = {E94B9BC6-2E64-4808-A710-BD8E3975D11D}
 		{FC12A768-F811-4CFD-B623-0AE749BA1354} = {A698F53C-280F-47E6-9910-F5B8DD3BF56B}
+		{5637E214-6C55-4085-8A38-F41E29EA4122} = {E94B9BC6-2E64-4808-A710-BD8E3975D11D}
+		{14B0C460-A362-47B4-A050-8B332DB79E2E} = {5637E214-6C55-4085-8A38-F41E29EA4122}
 	EndGlobalSection
 EndGlobal
