+++ tests/WarmCorners.Application.Tests.Unit/WarmCorners.Application.Tests.Unit.csproj
@@ -32,6 +32,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\WarmCorners.Application\WarmCorners.Application.csproj" />
+      <ProjectReference Include="..\WarmCorners.Tests.Common\WarmCorners.Tests.Common.csproj" />
     </ItemGroup>
 
 </Project>
