+++ tests/WarmCorners.Infrastructure.Tests.Unit/ServiceProviderBuilder.cs
@@ -5,6 +5,7 @@ using Serilog;
 using Serilog.Sinks.InMemory;
 using WarmCorners.Application.Common.Wrappers;
 using WarmCorners.Infrastructure.Wrappers;
+using WarmCorners.Tests.Common;
 
 namespace WarmCorners.Infrastructure.Tests.Unit;
 
@@ -14,11 +15,10 @@ public class ServiceProviderBuilder
 
     public ServiceProviderBuilder()
     {
-        this._services.AddInfrastructureServices();
+        this._services.AddInfrastructureServices()
+            .SetupInMemoryLogger();
 
         this.ReplaceWrappersWithMocks();
-
-        this.SetupInMemoryLogger();
     }
 
     public IServiceProvider Build() => this._services.BuildServiceProvider();
@@ -32,32 +32,5 @@ public class ServiceProviderBuilder
         var user32WrapperMock = this._services.ReplaceServiceWithMock<IUser32Wrapper>(ServiceLifetime.Transient);
         user32WrapperMock.Setup(u32W =>
             u32W.GetScreenResolution()).Returns((Testing.TestDisplaySize.Width, Testing.TestDisplaySize.Height));
-    }
-
-    private void SetupInMemoryLogger() =>
-        this._services.AddLogging(builder =>
-        {
-            builder.ClearProviders();
-
-            var logger = new LoggerConfiguration()
-                .WriteTo.InMemory()
-                .MinimumLevel.Verbose()
-                .CreateLogger();
-            builder.AddSerilog(logger);
-        });
-}
-
-public static class ServiceCollectionExtensions
-{
-    public static Mock<TService> ReplaceServiceWithMock<TService>(this IServiceCollection services, ServiceLifetime serviceLifetime)
-        where TService : class
-    {
-        var service = services.Single(sd => sd.ServiceType == typeof(TService));
-        services.Remove(service);
-        var replace = new Mock<TService>();
-        var serviceDescriptor = new ServiceDescriptor(typeof(TService), _ => replace.Object, serviceLifetime);
-        services.Add(serviceDescriptor);
-
-        return replace;
     }
 }
