+++ tests/WarmCorners.Service.Tests.Unit/ServiceProviderBuilder.cs
@@ -12,6 +12,7 @@ using SharpHook.Reactive;
 using WarmCorners.Application.Common.Services;
 using WarmCorners.Application.Common.Wrappers;
 using WarmCorners.Service.Configurations;
+using WarmCorners.Tests.Common;
 
 namespace WarmCorners.Service.Tests.Unit;
 
@@ -22,17 +23,37 @@ public class ServiceProviderBuilder
     public ServiceProviderBuilder()
     {
         this._services.AddPresentationServices(new ConfigurationBuilder().Build());
+        this._services.SetupInMemoryLogger();
+        this.AddSettings();
+
+        this.ReplaceServicesWithMocks();
 
         this.AddApplicationMocks();
         this.AddInfrastructureMocks();
-
-        this.AddSettings();
-
-        this.SetupReactiveGlobalHookMock();
-        this.SetupInMemoryLogger();
     }
 
     public IServiceProvider Build() => this._services.BuildServiceProvider();
+
+    private void AddSettings()
+    {
+        var triggerConfigurationOptionsMonitor = Mock.Of<IOptionsMonitor<TriggerConfiguration>>(om =>
+            om.CurrentValue == Testing.TriggerConfiguration);
+        this._services.AddSingleton(_ => triggerConfigurationOptionsMonitor);
+    }
+
+    private void ReplaceServicesWithMocks()
+    {
+        var reactiveGlobalHookMock = this._services.ReplaceServiceWithMock<IReactiveGlobalHook>(ServiceLifetime.Singleton);
+        reactiveGlobalHookMock
+            .Setup(rgh => rgh.MouseMoved)
+            .Returns(Testing.TestMouseMoved);
+        reactiveGlobalHookMock
+            .Setup(rgh => rgh.RunAsync())
+            .Returns(new Subject<System.Reactive.Unit>().AsObservable());
+        reactiveGlobalHookMock
+            .Setup(rgh => rgh.Dispose())
+            .Callback(() => { });
+    }
 
     private void AddApplicationMocks()
     {
@@ -48,53 +69,5 @@ public class ServiceProviderBuilder
         var schedulerWrapperMock = new Mock<ISchedulerWrapper>();
         schedulerWrapperMock.Setup(sw => sw.Default).Returns(Testing.TestScheduler);
         this._services.AddTransient(_ => schedulerWrapperMock.Object);
-    }
-
-    private void AddSettings()
-    {
-        var triggerConfigurationOptionsMonitor = Mock.Of<IOptionsMonitor<TriggerConfiguration>>(om =>
-            om.CurrentValue == Testing.TriggerConfiguration);
-        this._services.AddSingleton(_ => triggerConfigurationOptionsMonitor);
-    }
-
-    private void SetupReactiveGlobalHookMock()
-    {
-        var reactiveGlobalHookMock = this._services.ReplaceServiceWithMock<IReactiveGlobalHook>(ServiceLifetime.Singleton);
-        reactiveGlobalHookMock
-            .Setup(rgh => rgh.MouseMoved)
-            .Returns(Testing.TestMouseMoved);
-        reactiveGlobalHookMock
-            .Setup(rgh => rgh.RunAsync())
-            .Returns(new Subject<System.Reactive.Unit>().AsObservable());
-        reactiveGlobalHookMock
-            .Setup(rgh => rgh.Dispose())
-            .Callback(() => { });
-    }
-
-    private void SetupInMemoryLogger() =>
-        this._services.AddLogging(builder =>
-        {
-            builder.ClearProviders();
-
-            var logger = new LoggerConfiguration()
-                .WriteTo.InMemory()
-                .MinimumLevel.Verbose()
-                .CreateLogger();
-            builder.AddSerilog(logger);
-        });
-}
-
-public static class ServiceCollectionExtensions
-{
-    public static Mock<TService> ReplaceServiceWithMock<TService>(this IServiceCollection services, ServiceLifetime serviceLifetime)
-        where TService : class
-    {
-        var service = services.Single(sd => sd.ServiceType == typeof(TService));
-        services.Remove(service);
-        var replace = new Mock<TService>();
-        var serviceDescriptor = new ServiceDescriptor(typeof(TService), _ => replace.Object, serviceLifetime);
-        services.Add(serviceDescriptor);
-
-        return replace;
     }
 }
