+++ tests/WarmCorners.Tests.Integration/WarmCorners.Tests.Integration.csproj
@@ -36,6 +36,7 @@
     <ItemGroup>
       <ProjectReference Include="..\..\src\WarmCorners.Application\WarmCorners.Application.csproj" />
       <ProjectReference Include="..\..\src\WarmCorners.Infrastructure\WarmCorners.Infrastructure.csproj" />
+      <ProjectReference Include="..\WarmCorners.Tests.Common\WarmCorners.Tests.Common.csproj" />
     </ItemGroup>
 
     <ItemGroup>
